+++ sdk/eventhub/event-hubs/src/eventHubSender.ts
@@ -20,7 +20,8 @@ import {
   RetryOptions,
   defaultCancellableLock,
   retry,
-  translate
+  translate,
+  MessagingError
 } from "@azure/core-amqp";
 import { EventData, toRheaMessage } from "./eventData";
 import { ConnectionContext } from "./connectionContext";
@@ -339,21 +340,20 @@ export class EventHubSender extends LinkEntity {
     const timeoutInMs = getRetryAttemptTimeoutInMs(retryOptions);
     retryOptions.timeoutInMs = timeoutInMs;
 
-    const initStartTime = Date.now();
-    await this._createLinkIfNotOpen(options);
-    const timeTakenByInit = Date.now() - initStartTime;
-
     const sendEventPromise = async (): Promise<void> => {
+      const initStartTime = Date.now();
+      const sender = await this._createLinkIfNotOpen(options);
+      const timeTakenByInit = Date.now() - initStartTime;
       logger.verbose(
         "[%s] Sender '%s', credit: %d available: %d",
         this._context.connectionId,
         this.name,
-        this._sender!.credit,
-        this._sender!.session.outgoing.available()
+        sender.credit,
+        sender.session.outgoing.available()
       );
 
       let waitTimeForSendable = 1000;
-      if (!this._sender!.sendable() && timeoutInMs - timeTakenByInit > waitTimeForSendable) {
+      if (!sender.sendable() && timeoutInMs - timeTakenByInit > waitTimeForSendable) {
         logger.verbose(
           "%s Sender '%s', waiting for 1 second for sender to become sendable",
           this._context.connectionId,
@@ -366,14 +366,14 @@ export class EventHubSender extends LinkEntity {
           "%s Sender '%s' after waiting for a second, credit: %d available: %d",
           this._context.connectionId,
           this.name,
-          this._sender!.credit,
-          this._sender!.session?.outgoing?.available()
+          sender.credit,
+          sender.session?.outgoing?.available()
         );
       } else {
         waitTimeForSendable = 0;
       }
 
-      if (!this._sender!.sendable()) {
+      if (!sender.sendable()) {
         // let us retry to send the message after some time.
         const msg =
           `[${this._context.connectionId}] Sender "${this.name}", ` +
@@ -404,10 +404,9 @@ export class EventHubSender extends LinkEntity {
         throw translate(e);
       }
 
-      this._sender!.sendTimeoutInSeconds =
-        (timeoutInMs - timeTakenByInit - waitTimeForSendable) / 1000;
+      sender.sendTimeoutInSeconds = (timeoutInMs - timeTakenByInit - waitTimeForSendable) / 1000;
       try {
-        const delivery = await this._sender!.send(rheaMessage, undefined, 0x80013700, {
+        const delivery = await sender.send(rheaMessage, undefined, 0x80013700, {
           abortSignal
         });
         logger.info(
@@ -449,9 +448,9 @@ export class EventHubSender extends LinkEntity {
       retryOptions?: RetryOptions;
       abortSignal?: AbortSignalLike;
     } = {}
-  ): Promise<void> {
-    if (this.isOpen()) {
-      return;
+  ): Promise<AwaitableSender> {
+    if (this.isOpen() && this._sender) {
+      return this._sender;
     }
     const retryOptions = options.retryOptions || {};
     const timeoutInMs = getRetryAttemptTimeoutInMs(retryOptions);
@@ -459,7 +458,7 @@ export class EventHubSender extends LinkEntity {
     const senderOptions = this._createSenderOptions(timeoutInMs);
 
     const startTime = Date.now();
-    const createLinkPromise = async (): Promise<void> => {
+    const createLinkPromise = async (): Promise<AwaitableSender> => {
       return defaultCancellableLock.acquire(
         this.senderLock,
         () => {
@@ -475,7 +474,7 @@ export class EventHubSender extends LinkEntity {
       );
     };
 
-    const config: RetryConfig<void> = {
+    const config: RetryConfig<AwaitableSender> = {
       operation: createLinkPromise,
       connectionId: this._context.connectionId,
       operationType: RetryOperationType.senderLink,
@@ -484,7 +483,7 @@ export class EventHubSender extends LinkEntity {
     };
 
     try {
-      await retry<void>(config);
+      return await retry<AwaitableSender>(config);
     } catch (err) {
       const translatedError = translate(err);
       logger.warning(
@@ -500,6 +499,7 @@ export class EventHubSender extends LinkEntity {
 
   /**
    * Initializes the sender session on the connection.
+   * Should only be called from _createLinkIfNotOpen
    * @hidden
    */
   private async _init(
@@ -507,7 +507,7 @@ export class EventHubSender extends LinkEntity {
       abortSignal: AbortSignalLike | undefined;
       timeoutInMs: number;
     }
-  ): Promise<void> {
+  ): Promise<AwaitableSender> {
     try {
       if (!this.isOpen() && !this.isConnecting) {
         this.isConnecting = true;
@@ -526,7 +526,9 @@ export class EventHubSender extends LinkEntity {
           this.name
         );
 
-        this._sender = await this._context.connection.createAwaitableSender(options);
+        const sender = (this._sender = await this._context.connection.createAwaitableSender(
+          options
+        ));
         this.isConnecting = false;
         logger.verbose(
           "[%s] Sender '%s' created with sender options: %O",
@@ -539,8 +541,9 @@ export class EventHubSender extends LinkEntity {
         // It is possible for someone to close the sender and then start it again.
         // Thus make sure that the sender is present in the client cache.
         if (!this._context.senders[this.name]) this._context.senders[this.name] = this;
-        await this._ensureTokenRenewal();
-      } else {
+        this._ensureTokenRenewal();
+        return sender;
+      } else if (this._sender) {
         logger.verbose(
           "[%s] The sender '%s' with address '%s' is open -> %s and is connecting " +
             "-> %s. Hence not reconnecting.",
@@ -550,6 +553,9 @@ export class EventHubSender extends LinkEntity {
           this.isOpen(),
           this.isConnecting
         );
+        return this._sender;
+      } else {
+        throw new MessagingError("The sender link could not be opened.");
       }
     } catch (err) {
       this.isConnecting = false;